# action_server/tests/action_server_tests/test_server/test_server_full_openapi_flag.yml
components:
  schemas:
...
    ArtifactInfo:
      properties:
        name:
          title: Name
          type: string
        size_in_bytes:
          title: Size In Bytes
          type: integer
      required:
      - name
      - size_in_bytes
      title: ArtifactInfo
      type: object
    HTTPValidationError:
      properties:
        errors:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Errors
          type: array
      title: HTTPValidationError
      type: object
    Run:
      properties:
        action_id:
          title: Action Id
          type: string
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        id:
          title: Id
          type: string
        inputs:
          title: Inputs
          type: string
        numbered_id:
          title: Numbered Id
          type: integer
        relative_artifacts_dir:
          title: Relative Artifacts Dir
          type: string
        result:
          anyOf:
          - type: string
          - type: 'null'
          title: Result
        run_time:
          anyOf:
          - type: number
          - type: 'null'
          title: Run Time
        start_time:
          title: Start Time
          type: string
        status:
          title: Status
          type: integer
      required:
      - id
      - status
      - action_id
      - start_time
      - run_time
      - inputs
      - result
      - error_message
      - relative_artifacts_dir
      - numbered_id
      title: Run
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Robocorp Action Server
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/runs/{run_id}:
    get:
      operationId: show_run_api_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: ID for run
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Show Run
  /api/runs/{run_id}/artifacts:
    get:
      operationId: get_run_artifacts_api_runs__run_id__artifacts_get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: ID for run
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                description: 'Provides a list with the artifacts available

                  for a given run (i.e.: [{''name'': ''__action_server_output.txt'',
                  ''size_in_bytes'': 22}])

                  '
                items:
                  $ref: '#/components/schemas/ArtifactInfo'
                title: Response Get Run Artifacts Api Runs  Run Id  Artifacts Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Run Artifacts
...
servers:
- url: http://localhost:0